Chore(respec2html): Externalize fetch and write (closes #691)
